src/requirements.txt
requests>=2.31
click>=8.1.7
ruamel.yaml>=0.18.6
